test_graph v0.1.0 NeoSemantics.Extension

Module providing simple wrapper functions for the neosemantics library extension functions.

Functions

node_by_id(node_id, exclude_context \\ false)

Exports from Neo4j an LPG graph transformed into an RDF graph. The graph is centered on a node identified by node_id.

This returns the full context unless an optional boolean arg exclude_context is passed as true which will exclude context if present.

Examples

iex> NeoSemantics.Extension.node_by_id(1834)

node_by_uri(node_uri, exclude_context \\ false)

Exports from Neo4j an LPG graph transformed into an RDF graph. The graph is centered on a node identified by node_uri.

This returns the full context unless an optional boolean arg exclude_context is passed as true which will exclude context if present.

Examples

iex> NeoSemantics.Extension.node_by_uri("http://example.org/Elixir")

ping()

Pings the Neo4j service.

This returns a JSON confirmation that the service is up.

Examples

iex> NeoSemantics.Extension.ping
{"ping":"here!"}
:ok