test_graph v0.1.0 NeoSemantics.Inference

Module providing simple wrapper functions for the neosemantics library inference functions.

Functions

get_nodes_linked_to(conn, label)

Returns all nodes with label label or its sublabels.

get_nodes_with_label(conn, label)

Returns all nodes connected to node with label label or its subcategories.

get_rels(conn, rel)

Returns all outgoing relationships of type rel.