test_graph v0.1.0 NeoSemantics.Mapping

Module providing simple wrapper functions for the neosemantics library mapping functions.

Functions

add_common_schemas(conn)

Creates a references to a number of popular vocabularies including schema.org, Dublin Core, SKOS, OWL, etc.

add_mapping_to_schema(conn, node, element1, element2)

Creates a mapping for an element in the Neo4j DB schema to a vocabulary element.

add_schema(conn, uri, prefix)

Creates a reference to a vocabulary. Needed to define mappings.

drop_mapping(conn, element)

Returns an output text message indicating success/failure of the deletion.

drop_schema(conn, uri)

Deletes a vocabulary reference and all associated mappings.

list_mappings(conn, search_string \\ nil)

Returns a list with all the mappings.

list_schemas(conn, search_string \\ nil)

Returns all vocabulary references.