test_graph v0.1.0 TestGraph.RDF

Module providing a simple library for manipulating and querying RDF models.

Functions

read_graph()

Reads a default Turtle graph from the RDF graphs library.

Examples

iex> read_graph()
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> .

" ...

  file: "default.ttl",
  type: :rdf,
  uri: "file:///" <>... <> "/priv/rdf/graphs/default.ttl"
}

read_graph(list)

Reads a user Turtle graph from the RDF graphs library.

Examples

iex> read_graph("books.ttl")
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> .

" ...

  file: "books.ttl",
  type: :rdf,
  uri: "file:///" <>... <> "/priv/lpg/graphs/books.ttl"
}

write_graph(data)

Writes a Turtle graph to a temp file in the RDF graphs library.

Examples

iex> write_graph(data)
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> .

" ...

  file: "temp.ttl",
  type: :rdf,
  uri: "file:///" <>... <> "/priv/lpg/graphs/temp.ttl"
}

write_graph(data, list)

Writes a Turtle graph to a user file in the RDF graphs library.

Examples

iex> write_graph(data, file: "my.ttl")
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> .

" ...

  file: "my.ttl",
  type: :rdf,
  uri: "file:///" <>... <> "/priv/lpg/graphs/my.ttl"
}