test_super v0.1.0 TestSuper

Top-level module used in “Robust compute for RDF queries” post.

The post aims to demo how to manage fault tolerance in Elixir with supervision trees. It first reviews the process model for Erlang (Elixir) and then discusses the abstraction patterns that OTP provides for processes: agents, servers, supervisors, and tasks.

The post demos the GenServer, Supervisor, and DynamicSupervisor patterns.

Functions

genserver()

Constructor for new GenServer (with no supervision).

Returns process ID for GenServer.

Examples

iex> genserver
TestSuper.Server is starting ... #PID<0.224.0>
#PID<0.224.0>

genserver_d()

Constructor for new GenServer (with DynamicSupervisor).

Returns process ID for GenServer.

Examples

iex> genserver_d
TestSuper.Server is starting ... #PID<0.233.0>
#PID<0.233.0>

genserver_s()

Constructor for new GenServer (with Supervisor).

Returns process ID for GenServer.

Examples

iex> genserver_s
TestSuper.Server is starting ... #PID<0.229.0>
TestSuper.Supervisor is starting ... #PID<0.228.0>
#PID<0.229.0>

supervisor()

Constructor for new Supervisor (with GenServer).

Returns process ID for Supervisor.

Examples

iex> supervisor
TestSuper.Server is starting ... #PID<0.229.0>
TestSuper.Supervisor is starting ... #PID<0.228.0>
#PID<0.228.0>

supervisor_d()

Constructor for new DynamicSupervisor.

Returns process ID for DynamicSupervisor.

Examples

iex> supervisor_d
TestSuper.DynamicSupervisor is starting ... #PID<0.231.0>
#PID<0.231.0>