test_graph v0.1.0 TestGraph.LPG

Module providing a simple library for querying LPG models in a Neo4j instance via Cypher.

Functions

books()

Reads a Books graph from the graphs library.

Examples

iex> books().data
"//\n// create nodes\n//\nCREATE\n(book:Book {\n    iri: " <> ...

movies()

Reads a Movies graph from the graphs library.

Examples

iex> movies().data
"CREATE (TheMatrix:Movie {title:'The Matrix', released:1999," <> ...

parse(graphgist)

Parses a graphgist to return a Cypher graph.

Examples

iex> parse(read_graphgist())
"CREATE\n  (a:Person {name: 'Alice'}),\n  (b:Person {name: 'Bob'}),\n" <> ...

read_graph()

Reads a default Cypher graph from the LPG graphs library.

Examples

iex> read_graph()
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "default.cypher",
  type: :lpg,
  uri: "file:///" <>... <> "/priv/lpg/graphs/default.cypher"
}

read_graph(graph_file)

Reads a user Cypher graph from the LPG graphs library.

Examples

iex> read_graph("books.cypher")
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "books.cypher",
  type: :lpg,
  uri: "file:///" <>... <> "/priv/lpg/graphs/books.cypher"
}

read_graphgist()

Reads a default graphgist from the graphgists library.

Examples

iex> read_graphgist()
"= REPLACEME: TITLE OF YOUR GRAPHGIST\n:neo4j-version: 2.3.0\n:author:" <> ...

read_graphgist(graphgist_file)

Reads a user graphgist from the graphgists library.

Examples

iex> read_graphgist("template.adoc")
"= REPLACEME: TITLE OF YOUR GRAPHGIST\n:neo4j-version: 2.3.0\n:author:" <> ...

read_query()

Reads a default Cypher query from the LPG queries library.

Examples

iex> read_query()
"match (n) return n limit 1\n"

read_query(query_file)

Reads a named Cypher query from the LPG queries library.

Examples

iex> read_query("nodes.cypher")
"match (n) return n\n"

write_graph(data)

Writes a Turtle graph to a temp file in the LPG graphs library.

Examples

iex> data |> write_graph()
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "temp.cypher",
  type: :lpg,
  uri: "file:///" <>... <> "/priv/lpg/graphs/temp.cypher"
}

write_graph(data, graph_file)

Writes a Cypher graph to a user file in the LPG graphs library.

Examples

iex> data |> write_graph("my.cypher")
%TestGraph.Graph{
data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "my.cypher",
  type: :lpg,
  uri: "file:///" <>... <> "/priv/lpg/graphs/my.cypher"
}