test_graph v0.1.0 TestGraph.LPG

Module providing a simple library for querying LPG models in a Neo4j instance via Cypher.

Functions

books()

Reads a Books graph from the graphs library.

Examples

iex> books().data
"//\n// create nodes\n//\nCREATE\n(book:Book {\n    iri: " <> ...

movies()

Reads a Movies graph from the graphs library.

Examples

iex> movies().data
"CREATE (TheMatrix:Movie {title:'The Matrix', released:1999," <> ...

read_graph(graph_file \\ "default.cypher")

Reads a user Turtle graph from the RDF graphs library.

Examples

iex> read_graph()
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "default.cypher",
  path:  ... <> "/test_graph/priv/lpg/graphs/default.cypher",
  type: :lpg,
  uri: "file:///" <> ... <> "/test_graph/priv/lpg/graphs/default.cypher"
}

iex> read_graph("books.cypher")
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "books.cypher",
  path:  ... <> "/test_graph/priv/lpg/graphs/books.cypher",
  type: :lpg,
  uri: "file:///" <> ... <> "/test_graph/priv/lpg/graphs/books.cypher"
}

read_query(query_file \\ "default.cypher")

Reads a Cypher query from the LPG queries library.

Examples

iex> read_query()
%TestGraph.Query{
  data: "match (n) return n\n"
  file: "nodes.cypher",
  path:  ... <> "/test_graph/priv/lpg/queries/nodes.cypher",
  type: :lpg,
  uri: "file:///" <> ... <> "/test_graph/priv/lpg/queries/nodes.cypher"
}

iex> read_query("nodes.cypher")
%TestGraph.Query{
  data: "match (n) return n\n"
  file: "nodes.cypher",
  path:  ... <> "/test_graph/priv/lpg/queries/nodes.cypher",
  type: :lpg,
  uri: "file:///" <> ... <> "/test_graph/priv/lpg/queries/nodes.cypher"
}

write_graph(graph_data, graph_file \\ "temp.cypher")

Writes a Turtle graph to a user file in the RDF graphs library.

Examples

iex> data |> write_graph("my.cypher")
%TestGraph.Graph{
  data: "//\n// create nodes\n//\nCREATE\n(book:Book {\n" <> ...
  file: "my.cypher",
  path:  ... <> "/test_graph/priv/lpg/graphs/my.cypher",
  type: :lpg,
  uri: "file:///" <>... <> "/test_graph/priv/lpg/graphs/my.cypher"
}

write_query(query_data, query_file \\ "temp.cypher")

Writes a Cypher query to a file in the LPG queries library.

Examples

iex> write_query("my.cypher")
%TestGraph.Query{
  data: "match (n) return n\n"
  file: "my.cypher",
  path:  ... <> "/test_graph/priv/lpg/queries/my.cypher",
  type: :lpg,
  uri: "file:///" <> ... <> "/test_graph/priv/lpg/queries/my.cypher"
}