test_graph v0.1.0 TestGraph.RDF

Module providing a simple library for manipulating and querying RDF models.

Functions

book()

list_graphs()

Lists Turtle graphs in the RDF graphs library.

Examples

iex> list_graphs()
["books.ttl", "urn_isbn_978-1-68050-252-7.ttl",
 "http___dbpedia.org_resource_London.ttl", "london100.ttl", "london.ttl",
 "nobelprizes.ttl", "bibo.ttl", "tony.ttl", "temp.ttl",
 "http___example.org_Elixir.ttl", "elixir.ttl", "default.ttl", "cypher.ttl",
 "neo4j.ttl", "hello.ttl"]

list_queries()

Lists SPARQL queries in the RDF queries library.

Examples

iex> list_queries()
["cypher.rq", "london.rq", "elixir.rq", "default.rq", "neo4j.rq",
 "triples_by_uri.rq", "triples.rq", "hello.rq"]

read_graph(graph_file \\ "default.ttl")

Reads a Turtle graph from the RDF graphs library.

Examples

iex> read_graph()
%TestGraph.Graph{
  data: "<http://dbpedia.org/resource/Hello_World>\n" <> ...
  file: "default.ttl",
  path:  ... <> "/test_graph/priv/rdf/graphs/default.ttl",
  type: :rdf,
  uri: "file:///" <> ... <> "/test_graph/priv/rdf/graphs/default.ttl"
}

iex> read_graph("books.ttl")
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> \n" <> ...
  file: "books.ttl",
  path:  ... <> "/test_graph/priv/rdf/graphs/books.ttl",
  type: :rdf,
  uri: "file:///" <> ... <> "/test_graph/priv/rdf/graphs/books.ttl"
}

read_query(query_file \\ "default.rq")

Reads a SPARQL query from the RDF queries library.

Examples

iex> read_query()
%TestGraph.Query{
  data: "construct { ?s ?p ?o } where { " <> ...
  file: "books.rq",
  path:  ... <> "/test_graph/priv/rdf/queries/books.rq",
  type: :rdf,
  uri: "file:///" <> ... <> "/test_graph/priv/rdf/queries/books.rq"
}

iex> read_query("books.rq")
%TestGraph.Query{
  data: "construct { ?s ?p ?o } where { " <> ...
  file: "books.rq",
  path:  ... <> "/test_graph/priv/rdf/queries/books.rq",
  type: :rdf,
  uri: "file:///" <> ... <> "/test_graph/priv/rdf/queries/books.rq"
}

write_graph(graph_data, graph_file \\ "temp.ttl")

Writes a Turtle graph to a file in the RDF graphs library.

Examples

iex> data |> write_graph("my.ttl")
%TestGraph.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> \n" <> ...
  file: "my.ttl",
  path:  ... <> "/test_graph/priv/rdf/graphs/my.ttl",
  type: :rdf,
  uri: "file:///" <>... <> "/test_graph/priv/rdf/graphs/my.ttl"
}

write_query(query_data, query_file \\ "temp.rq")

Writes a SPARQL query to a file in the RDF queries library.

Examples

iex> write_query("my.rq")
%TestGraph.Query{
  data: "construct { ?s ?p ?o } where { " <> ...
  file: "my.rq",
  path:  ... <> "/test_graph/priv/rdf/queries/my.rq",
  type: :rdf,
  uri: "file:///" <> ... <> "/test_graph/priv/rdf/queries/my.rq"
}