test_neo4j v0.1.0 TestNeo4j

Top-level module used in "Property graphs and Elixir" post.

Functions

books()

Reads a Books graph from the graphs library.

clear(conn)

Deletes all nodes and relationships in database.

init()

Opens up a Bolt database connection with the app config.

movies()

Reads a Movies graph from the graphs library.

node1(conn)

Queries database for one node.

node1_and_relationships(conn)

Queries database for one node and relationships.

nodes(conn)

Queries database for all nodes.

nodes_and_relationships(conn)

Queries database for all nodes and relationships.

parse(graphgist)

Parses a graphgist to return a Cypher graph.

path1(conn)

Queries database for one path.

paths(conn)

Queries database for all paths.

read_graph()

Reads a default Cypher graph from the graphs library.

read_graph(graph_file)

Reads a user Cypher graph from the graphs library.

read_graphgist()

Reads a default graphgist from the graphgists library.

read_graphgist(graphgist_file)

Reads a user graphgist from the graphgists library.

read_query()

Reads a default Cypher query from the queries library.

read_query(query_file)

Reads a named Cypher query from the queries library.

relationship1(conn)

Queries database for one relationship.

relationships(conn)

Queries database for all relationships.

reset(conn)

Deletes all nodes and relationships in database.

test(conn)

Counts nodes, relationships and paths in database.